LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/drawing-rectangles-to-the-canvas/2d.clearRect.transform-expected.txt

2d.clearRect.transform

clearRect is affected by transforms


PASS clearRect is affected by transforms 
